Services/chartIQ-python/pyproject.toml
[tool.poetry]
name = "chartiq"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

# System
[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.8.2"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
openai = "1.43.0"
python-dotenv = "^1.0.1"
jsbeautifier = "^1.15.1"
aiofiles = "^24.1.0"
jinja2 = "^3.1.4"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# Configs
## Ruff
[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py311"
src = ["src"]

# Linter settings.
[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
lines-after-imports = 2
lines-between-types = 1

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"


## Local development
[tool.poetry.group.dev]
optional = true


[tool.poetry.group.dev.dependencies]
ruff = "0.5.2"
mypy = "1.11.2"


## MyPy
[tool.mypy]
plugins = ["pydantic.mypy"]
disable_error_code = ["no-redef", "func-returns-value"]
strict = false
check_untyped_defs = true
strict_optional = true
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
